Rename ReadonlyLazyListCompactTest to LazyDeserializationTest and add ReadResultSet test
