+++ src/proxy/ringbuffer/LazyReadResultSet.ts
@@ -57,8 +57,9 @@ export class LazyReadResultSet<T> implements ReadResultSet<T> {
                 obj = this.serializationService.toObject(dataOrObject);
             } catch (e) {
                 if (e instanceof SchemaNotFoundError) {
-                    throw new HazelcastSerializationError(e.message, e.cause, e.serverStackTrace);
+                    throw new HazelcastSerializationError(e.message, e, e.serverStackTrace);
                 }
+                throw e;
             }
             this.items[index] = obj;
             return obj;
